+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - aria-hidden attribute to all Divider components
+- Cookiebot declaration view
 
 ### Fixed